refactor(cli): 拆分 ts_bridge 快照同步与事件投影管线
- apply_snapshot / apply_bridge_events 迁入 sync/
- 生命周期与 call_bridge 迁入 runtime/lifecycle.rs
