+++ apps/cli/src/ts_bridge/runtime/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod lifecycle;
