+++ apps/cli/src/ts_bridge/sync/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod events;
+pub(crate) mod snapshot;
+pub(crate) mod subagent_push;
